update: rename GetClubServiceImpl to ClubService and update method names
