+++ src/main/java/team/themoment/imi/domain/club/service/ClubService.java
@@ -1,22 +1,20 @@
-package team.themoment.imi.domain.club.service.impl;
+package team.themoment.imi.domain.club.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import team.themoment.imi.domain.club.data.response.ClubListResDto;
-import team.themoment.imi.domain.club.enitity.Club;
+import team.themoment.imi.domain.club.entity.Club;
 import team.themoment.imi.domain.club.repository.ClubJpaRepository;
-import team.themoment.imi.domain.club.service.GetClubService;
 
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class GetClubServiceImpl implements GetClubService {
+public class ClubService {
 
     private final ClubJpaRepository clubJpaRepository;
 
-    @Override
-    public ClubListResDto execute() {
+    public ClubListResDto getAllClubs() {
         List<Club> clubs = clubJpaRepository.findAll();
         return new ClubListResDto(clubs.size(), clubs);
     }
