--- src/main/java/team/themoment/imi/domain/club/service/GetClubService.java
@@ -1,7 +1,0 @@
-package team.themoment.imi.domain.club.service;
-
-import team.themoment.imi.domain.club.data.response.ClubListResDto;
-
-public interface GetClubService {
-    ClubListResDto execute();
-}
